fix(phaseflux): 16-pulse mask coverage + pitch-window boundary hold
- _pulseFired widened uint8_t -> uint16_t so all 16 scheduled pulses can be
  marked fired (bits 8-15 were silently dropped).
- Gate-mode pitch windows no longer drop pitch pulses in hidden bands — they
  clamp to the nearest visible boundary; always-mode uses the same hold.
- Tests for 16-pulse mask coverage and pitch-window boundary hold.

+++ src/apps/sequencer/engine/PhaseFluxTrackEngine.h
@@ -112,7 +112,7 @@ private:
     uint32_t _prevCycleTick = 0;
 
     // Per-stage-visit pulse-fire memory (bit i = pulse i fired).
-    uint8_t _pulseFired = 0;
+    uint16_t _pulseFired = 0;
 
     // Derived position.
     int _slotIdx = 0;
